emb-bench/requirements.txt
numpy==2.1.3
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
qdrant-client==1.12.1
fastembed==0.6.0
sentence-transformers==3.3.1
scikit-learn==1.6.1
